Added info about race detection and tracing
